refactor jsf subsystem transformers
